refactor: made a more compact version of the components
