+++ lib/pearl_web/live/landing/challenges_live/index.html.heex
@@ -9,18 +9,19 @@
 
 <div class="w-full mx-auto px-6.5 md:px-16 mb-20 pt-8">
   <div class="mt-4 md:mt-28 mb-24 grid grid-cols-1 gap-12 xl:grid-cols-[400px_1fr]">
-    <.desktop_sidebar
+    <.challenges_list
       challenges={@challenges}
       selected_challenge_id={@selected_challenge.id}
+      view_mode="desktop"
     />
 
-    <.challenge_card challenge={@selected_challenge} />
+    <.challenge_detail challenge={@selected_challenge} view_mode="desktop" />
 
     <div class="xl:hidden">
       <%= if assigns[:mobile_selected_challenge] do %>
-        <.mobile_detail challenge={@mobile_selected_challenge} />
+        <.challenge_detail challenge={@mobile_selected_challenge} view_mode="mobile" />
       <% else %>
-        <.mobile_list challenges={@challenges} />
+        <.challenges_list challenges={@challenges} view_mode="mobile" />
       <% end %>
     </div>
   </div>
